# specs/005-workspace-completion/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "GradePreview": {
        "additionalProperties": false,
        "properties": {
          "final_score": {
            "title": "Final Score",
            "type": "number"
          },
          "pass_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pass Score"
          },
          "penalty": {
            "title": "Penalty",
            "type": "number"
          },
          "penalty_days": {
            "minimum": 0.0,
            "title": "Penalty Days",
            "type": "integer"
          },
          "penalty_rate": {
            "title": "Penalty Rate",
            "type": "number"
          },
          "policy_revision": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Revision"
          },
          "raw_score": {
            "title": "Raw Score",
            "type": "number"
          }
        },
        "required": [
          "raw_score",
          "penalty_days",
          "penalty_rate",
          "penalty",
          "final_score",
          "pass_score",
          "policy_revision"
        ],
        "title": "GradePreview",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PublishedCriterionView": {
        "additionalProperties": false,
        "properties": {
          "max_points": {
            "title": "Max Points",
            "type": "number"
          },
          "points": {
            "title": "Points",
            "type": "number"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "points",
          "max_points",
          "reason"
        ],
        "title": "PublishedCriterionView",
        "type": "object"
      },
...
      "StudentReviewView": {
        "additionalProperties": false,
        "properties": {
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/PublishedCriterionView"
            },
            "title": "Criteria",
            "type": "array"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "decision_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Reason"
          },
          "feedback": {
            "title": "Feedback",
            "type": "string"
          },
          "grade": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GradePreview"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "iteration_id": {
            "format": "uuid",
            "title": "Iteration Id",
            "type": "string"
          },
          "published_at": {
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "revision_deadline": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision Deadline"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "submission_version_id": {
            "format": "uuid",
            "title": "Submission Version Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "iteration_id",
          "submission_version_id",
          "published_at",
          "score",
          "feedback",
          "decision",
          "revision_deadline",
          "criteria"
        ],
        "title": "StudentReviewView",
        "type": "object"
      },
      "StudentSubmissionView": {
        "additionalProperties": false,
        "properties": {
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/SubmissionAttemptView"
            },
            "title": "Attempts",
            "type": "array"
          },
          "course_run_id": {
            "format": "uuid",
            "title": "Course Run Id",
            "type": "string"
          },
          "current_publication_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Publication Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "publication_id": {
            "format": "uuid",
            "title": "Publication Id",
            "type": "string"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/StudentReviewView"
            },
            "title": "Reviews",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "publication_id",
          "course_run_id",
          "attempts",
          "reviews",
          "current_publication_id"
        ],
        "title": "StudentSubmissionView",
        "type": "object"
      },
      "SubmissionAttemptView": {
        "additionalProperties": false,
        "properties": {
          "artifact_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Id"
          },
          "capture_operation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capture Operation Id"
          },
          "comment": {
            "title": "Comment",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "sequence": {
            "minimum": 0.0,
            "title": "Sequence",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "submitted_at": {
            "format": "date-time",
            "title": "Submitted At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "comment",
          "sequence",
          "submitted_at",
          "status",
          "artifact_id",
          "capture_operation_id"
        ],
        "title": "SubmissionAttemptView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Review Workspace Extension",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v2/submissions/{identity}": {
      "get": {
        "operationId": "get_student_submission_api_v2_submissions__identity__get",
        "parameters": [
          {
            "in": "path",
            "name": "identity",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Identity",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentSubmissionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Student Submission",
        "tags": [
          "workspace-v2"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}